Add NIST P-256 support
Add a flag `--ec256` to allow the use of NIST P-256 curve in key pair
generation. The command

    pivit --generate --ec256

will create a key pair using the P-256 curve, while

    pivit --generate

still generates a key pair using the P-384 curve.

The use of a boolean flag as the curve indicator isn't the most
flexible, but is likely the minimum to get this feature supported.

Update README.md to explain the new `--ec256` flag.