Add command line flags to control PIN and touch policies
When generating a hardware backed key using the PIV applet, it's possible to set the key access policy.
The PIN policy controls if/when to prompt the user for a PIN when accessing the generated key.
The touch policy controls if/when to prompt the user to touch the hardware to access the generated key.

By default, these policies are set to PIN=never and touch=always.
We believe that these policies provide less friction to users while still maintaining reasonable security settings.

For more details on the PIV PIN and touch policies, see Yubico's [user's manual](https://docs.yubico.com/yesdk/users-manual/application-piv/pin-touch-policies.html)

Signed-off-by: Yoav Amit <yoav@squareup.com>