Added "cached" option when setting the touch policy
This option is mostly supported by Yubico hardware with PIV applets, but could still be useful.

The meaning of setting this option is that once a connection to the PIV hardware is opened, it'll only prompt the user to touch the key 15 minutes after the first time it was touched since the connection was opened.
This option is mostly helpful when you need to access the private key inside the PIV applet many times in a single connection.