Store PIV management key in Yubikey metadata (PIN protected)
Until now, Pivit used to derive the 24 byte long management key by perform a SHA256 checksum of the PIN.
It was done so Pivit could easily be able to derive the management key without user interaction.

A better and more standard (and slightly more secure) way of doing this, is to create a random management key and then store it in the PIV applet's "metadata" section which is PIN protected.