Move certificate parameters to a dedicated struct
When generating a certificate or a certificate signing request, instead of getting the certificate subject information directly from environment variables, they'll be passed to the `GenerateCertificate` function in a dedicated struct.
The environment variables will still be used by the pivit CLI executable to populate the `CertificateParameters` struct when calling `pivit --generate`.